Add: findable module
