+++ lib/stat_tracker.rb
@@ -1,10 +1,13 @@
 require 'simplecov'
 SimpleCov.start
+# require_relative './spec_helper'
+
 require_relative './game.rb'
 require_relative './league.rb'
 require_relative './team.rb'
 require_relative './season.rb'
-require 'pry'
+
+# require 'pry'
 require 'csv'
 
 class StatTracker
